Make sequence 6 digits and fix to migration sequence count off
This fixes a problem with the sequence autogeneration during the
migration. Originally it was starting from the "need to do" number,
which was incorrect. Now it counts the total number of components for a
collection, then figures out those that don't need new cuids and starts
the sequence from that number. The sequence is then set to the total
number of components, so new records should start from the correct
number.